Added context menu item for links!
Right-clicking on link now shows "Open Link with Zoom.it".
TODO: right-clicking on an *image* link shows both options, need to figure out
how to fix this. It's confusing and it also combines both into a submenu.

Also updated the manifest:
+ Improved description; now explicit about what this extension adds.
+ Upped the version from 1.0 to 1.1.
+ Fixed 3-space tabs to 4 spaces.

Also minor code cleanup, and improved wording ("with" Zoom.it now).